backend.chore: refactor configuration model into ParcelManagement.Core.Model.Configuration
