+++ backend/src/ParcelManagement.Core/Model/Configuration/RedisSettings.cs
@@ -1,4 +1,4 @@
-namespace ParcelManagement.Core.Model
+namespace ParcelManagement.Core.Model.Configuration
 {
     public class RedisSettings
     {
